Logik Teil neuer Name. Logik funktioniert jetzt mit allen .fasta file durch try-except Block. Anpassung der GUI Logik auf separate Methoden
